Invisibility fixes
Found some issues with getting the player to be fully opaque again when using invisible. Seems the trigger for character model visibility is tied to the name plate visibility. Talking to an NPC would fix it, but that's not ideal. For now, the name plate will flash if buff ID 69 is dropped. There may be some other buff IDs that might need to get checked for, but trying this for now.

Also added a similar brief nameplate flash when the addon is loaded, so reloading the app properly resets itself.

Name visibility is now reset when unloading the app.